+++ Changelog.md
@@ -5,6 +5,7 @@
 
 - Added the missing order status `processing` enum icon
 - Fixed the invalid icons on carrier and shipping method forms
+- Added modal to display configuration JSON of order items
 
 ## 4.1.0
 ##### 2024-07-11